Refactored ranking
